# [WIP] Faster architecture-specific minifier