--- a/benchmark/minifiercompetition.cpp
+++ b/benchmark/minifiercompetition.cpp
@@ -108,6 +108,9 @@ int main(int argc, char *argv[]) {
   BEST_TIME("json_minify", simdjson::json_minify(cbuffer, p.size(), cbuffer),
             outlength, memcpy(buffer, p.data(), p.size()), repeat, volume,
             !just_data);
+  BEST_TIME("simdjson->minify", (simdjson::active_implementation->minify(cbuffer, p.size(), cbuffer, outlength), outlength),
+            outlength, memcpy(buffer, p.data(), p.size()), repeat, volume,
+            !just_data);
   printf("minisize = %zu, original size = %zu  (minified down to %.2f percent "
          "of original) \n",
          outlength, p.size(), outlength * 100.0 / p.size());
@@ -171,6 +174,7 @@ int main(int argc, char *argv[]) {
                                  automated_reallocation),
             simdjson::SUCCESS, memcpy(buffer, mini_buffer, p.size()), repeat, volume,
             !just_data);
+
   free(buffer);
   free(ast_buffer);
   free(mini_buffer);