--- a/src/arm64/bitmanipulation.h
+++ b/src/arm64/bitmanipulation.h
@@ -48,7 +48,7 @@ really_inline int leading_zeroes(uint64_t input_num) {
 }
 
 /* result might be undefined when input_num is zero */
-really_inline int hamming(uint64_t input_num) {
+really_inline int count_ones(uint64_t input_num) {
    return vaddv_u8(vcnt_u8((uint8x8_t)input_num));
 }