--- a/src/generic/json_scanner.h
+++ b/src/generic/json_scanner.h
@@ -5,23 +5,33 @@ namespace stage1 {
  */
 struct json_block {
 public:
-  // the start of structurals that are not inside strings
+  /** The start of structurals */
   really_inline uint64_t structural_start() { return potential_structural_start() & ~_string.string_tail(); }
+  /** All JSON whitespace (i.e. not in a string) */
+  really_inline uint64_t whitespace() { return non_quote_outside_string(_characters.whitespace()); }
 
-  // operators plus scalar starts like 123, true and "abc"
-  really_inline uint64_t potential_structural_start() { return _characters.op() | potential_scalar_start(); }
-  // the start of non-operator runs, like 123, true and "abc"
-  really_inline uint64_t potential_scalar_start() { return _characters.scalar() & ~follows_potential_scalar(); }
-  // whether the given character is immediately after a non-operator like 123, true or "
-  really_inline uint64_t follows_potential_scalar() { return _follows_potential_scalar; }
-  // Return a mask of whether the given characters are inside a string (only works on non-quotes)
+  // Helpers
+
+  /** Whether the given characters are inside a string (only works on non-quotes) */
   really_inline uint64_t non_quote_inside_string(uint64_t mask) { return _string.non_quote_inside_string(mask); }
+  /** Whether the given characters are outside a string (only works on non-quotes) */
+  really_inline uint64_t non_quote_outside_string(uint64_t mask) { return _string.non_quote_outside_string(mask); }
+
   // string and escape characters
   json_string_block _string;
   // whitespace, operators, scalars
   json_character_block _characters;
   // whether the previous character was a scalar
   uint64_t _follows_potential_scalar;
+private:
+  // Potential structurals (i.e. disregarding strings)
+
+  /** operators plus scalar starts like 123, true and "abc" */
+  really_inline uint64_t potential_structural_start() { return _characters.op() | potential_scalar_start(); }
+  /** the start of non-operator runs, like 123, true and "abc" */
+  really_inline uint64_t potential_scalar_start() { return _characters.scalar() & ~follows_potential_scalar(); }
+  /** whether the given character is immediately after a non-operator like 123, true or " */
+  really_inline uint64_t follows_potential_scalar() { return _follows_potential_scalar; }
 };
 
 /**