--- a/src/haswell/simd.h
+++ b/src/haswell/simd.h
@@ -302,7 +302,6 @@ namespace simdjson::haswell::simd {
       const simd8<T> mask = simd8<T>::splat(m);
       return this->map( [&](auto a) { return a <= mask; } ).to_bitmask();
     }
-
   }; // struct simd8x64<T>
 
 } // namespace simdjson::haswell::simd