--- a/src/westmere/simd.h
+++ b/src/westmere/simd.h
@@ -302,7 +302,6 @@ namespace simdjson::westmere::simd {
       const simd8<T> mask = simd8<T>::splat(m);
       return this->map( [&](auto a) { return a <= mask; } ).to_bitmask();
     }
-
   }; // struct simd8x64<T>
 
 } // namespace simdjson::westmere::simd