This is a minifier using the actual stage 1 code, taking advantage of the new structural scanner. It's faster than json_minify, by about 2x. Interestingly, it's slower than parsing the file into a DOM. I'm pretty sure there are optimizations that can be done here, however; it's somewhat naive, grabbing the mask of non-whitespace (taking strings into account) and using memcpy() on each non-whitespace region. I haven't tried any variant optimizations at all.

```
make parsingcompetition && ./parsingcompetition jsonexamples/twitter.json

despacing with RapidJSON                        :  9.156 cycles per input byte (best)  10.995 cycles per input byte (avg) 
despacing with RapidJSON Insitu                 :  8.696 cycles per input byte (best)  9.148 cycles per input byte (avg) 
json_minify                                     :    2.182 cycles per input byte (best)    2.326 cycles per input byte (avg)    0.910 GB/s (error margin: 0.058 GB/s)
simdjson->minify                                :    1.152 cycles per input byte (best)    1.224 cycles per input byte (avg)    1.719 GB/s (error margin: 0.104 GB/s)
minisize = 466906, original size = 631515  (minified down to 73.93 percent of original) 
RapidJSON Insitu orig                           :    3.498 cycles per input byte (best)    3.867 cycles per input byte (avg)    0.568 GB/s (error margin: 0.055 GB/s)
RapidJSON Insitu despaced                       :    3.011 cycles per input byte (best)    3.300 cycles per input byte (avg)    0.660 GB/s (error margin: 0.058 GB/s)
sajson orig                                     :    1.770 cycles per input byte (best)    1.910 cycles per input byte (avg)    1.121 GB/s (error margin: 0.083 GB/s)
sajson despaced                                 :    1.427 cycles per input byte (best)    1.617 cycles per input byte (avg)    1.389 GB/s (error margin: 0.165 GB/s)
simdjson orig                                   :    0.712 cycles per input byte (best)    0.856 cycles per input byte (avg)    2.772 GB/s (error margin: 0.468 GB/s)
simdjson despaced                               :    0.665 cycles per input byte (best)    1.169 cycles per input byte (avg)    2.965 GB/s (error margin: 1.279 GB/s)
```

I'm only despacing, is that right? You could make JSON smaller by canonicalizing it--i.e. by parsing \uXXXX, and by replacing 0.000 with 0.0 or even 0--but I don't believe we're doing that, and I'm not sure anyone else does ...