@lemire 2020-03-19
Minifying can be surprisingly difficult and expensive. The proper way is to vectorize it...

The proper way on x64 is to rely on the following magic:

```C++
// Write up to 16 bytes, only the bytes corresponding to a 1-bit are written
// out. credit: Anime Tosho
static __m128i skinnycleanm128(__m128i x, int mask) {
  int mask1 = mask & 0xFF;
  int mask2 = (mask >> 8) & 0xFF;
  __m128i shufmask = _mm_castps_si128(
      _mm_loadh_pi(_mm_castsi128_ps(_mm_loadl_epi64(
                       (const __m128i *)(thintable_epi8 + mask1))),
                   (const __m64 *)(thintable_epi8 + mask2)));
  shufmask =
      _mm_add_epi8(shufmask, _mm_set_epi32(0x08080808, 0x08080808, 0, 0));
  __m128i pruned = _mm_shuffle_epi8(x, shufmask);
  intptr_t popx2 = BitsSetTable256mul2[mask1];
  __m128i compactmask =
      _mm_loadu_si128((const __m128i *)(pshufb_combine_table + popx2 * 8));
  return _mm_shuffle_epi8(pruned, compactmask);
}
```

(The thintable_epi8 is in our code base.)

It looks ugly and so forth, but it is really just two calls to _mm_shuffle_epi8.

You first pack (remove spaces) in two 8-byte subwords. Then you slide them together. This is complicated, a bit, but the end result is that you can do all of the work with a small lookup table.

I think I can port it on top of your PR if you stand still long enough.