@jkeiser 2020-03-19
Without the branches writing the output, I bet we could pull 5GB/s (around the time spent in current state 1).