@lemire 2020-03-19
I'll work something out.