@lemire 2020-03-19
I should be able to have a prototype ready today if people leave me alone for an hour or so.