@lemire 2020-03-20
I am hoping that this is made obsolete by https://github.com/lemire/simdjson/pull/568